+++ src/schema.mli
@@ -14,7 +14,6 @@ val default : loc:Warnings.loc -> Ppxlib.expression -> Ppxlib.expression -> Ppxl
 val description : loc:Warnings.loc -> string -> Ppxlib.expression -> Ppxlib.expression
 val variants :
   loc:Warnings.loc ->
-  ?as_string:bool ->
   ?compact_variants:bool ->
   [< `Inherit of Ppxlib.expression | `Tag of string * Ppxlib.expression list * string option ] list ->
   Ppxlib.expression