+++ test/shared/generate_schemas_cases.ml
@@ -10,13 +10,9 @@ let schemas =
     Ppx_deriving_jsonschema_runtime.json_schema Mod1.Mod2.m_2_jsonschema;
     Ppx_deriving_jsonschema_runtime.json_schema with_modules_jsonschema;
     Ppx_deriving_jsonschema_runtime.json_schema kind_jsonschema;
-    Ppx_deriving_jsonschema_runtime.json_schema kind_as_string_jsonschema;
     Ppx_deriving_jsonschema_runtime.json_schema poly_kind_jsonschema;
-    Ppx_deriving_jsonschema_runtime.json_schema poly_kind_as_string_jsonschema;
     Ppx_deriving_jsonschema_runtime.json_schema poly_kind_with_payload_jsonschema;
-    Ppx_deriving_jsonschema_runtime.json_schema poly_kind_with_payload_as_string_jsonschema;
     Ppx_deriving_jsonschema_runtime.json_schema poly_inherit_jsonschema;
-    Ppx_deriving_jsonschema_runtime.json_schema poly_inherit_as_string_jsonschema;
     Ppx_deriving_jsonschema_runtime.json_schema event_jsonschema;
     Ppx_deriving_jsonschema_runtime.json_schema recursive_record_jsonschema;
     Ppx_deriving_jsonschema_runtime.json_schema recursive_variant_jsonschema;
@@ -39,7 +35,6 @@ let schemas =
     Ppx_deriving_jsonschema_runtime.json_schema numbers_jsonschema;
     Ppx_deriving_jsonschema_runtime.json_schema opt_jsonschema;
     Ppx_deriving_jsonschema_runtime.json_schema using_m_jsonschema;
-    Ppx_deriving_jsonschema_runtime.json_schema (poly2_jsonschema int_jsonschema);
     Ppx_deriving_jsonschema_runtime.json_schema tuple_with_variant_jsonschema;
     Ppx_deriving_jsonschema_runtime.json_schema ~id:"https://ahrefs.com/schemas/player_scores" ~title:"Player scores"
       ~description:"Object representing player scores"
@@ -48,11 +43,8 @@ let schemas =
     Ppx_deriving_jsonschema_runtime.json_schema t_jsonschema;
     Ppx_deriving_jsonschema_runtime.json_schema ~definitions:[ "shared_address", address_jsonschema ] tt_jsonschema;
     Ppx_deriving_jsonschema_runtime.json_schema c_jsonschema;
-    Ppx_deriving_jsonschema_runtime.json_schema variant_inline_record_jsonschema;
     Ppx_deriving_jsonschema_runtime.json_schema inline_record_with_extra_fields_jsonschema;
-    Ppx_deriving_jsonschema_runtime.json_schema variant_with_payload_jsonschema;
     Ppx_deriving_jsonschema_runtime.json_schema t1_jsonschema;
-    Ppx_deriving_jsonschema_runtime.json_schema t2_jsonschema;
     Ppx_deriving_jsonschema_runtime.json_schema t3_jsonschema;
     Ppx_deriving_jsonschema_runtime.json_schema t4_jsonschema;
     Ppx_deriving_jsonschema_runtime.json_schema t5_jsonschema;
@@ -71,13 +63,11 @@ let schemas =
     Ppx_deriving_jsonschema_runtime.json_schema (param_list_jsonschema string_jsonschema);
     Ppx_deriving_jsonschema_runtime.json_schema (either_jsonschema int_jsonschema string_jsonschema);
     Ppx_deriving_jsonschema_runtime.json_schema (either_alias_jsonschema int_jsonschema string_jsonschema);
-    Ppx_deriving_jsonschema_runtime.json_schema (direction_jsonschema int_jsonschema string_jsonschema);
     Ppx_deriving_jsonschema_runtime.json_schema tool_params_jsonschema;
     Ppx_deriving_jsonschema_runtime.json_schema described_record_jsonschema;
     Ppx_deriving_jsonschema_runtime.json_schema with_key_and_desc_jsonschema;
     Ppx_deriving_jsonschema_runtime.json_schema described_variant_jsonschema;
     Ppx_deriving_jsonschema_runtime.json_schema described_variant_inline_record_jsonschema;
-    Ppx_deriving_jsonschema_runtime.json_schema described_variant_string_jsonschema;
     Ppx_deriving_jsonschema_runtime.json_schema doc_comment_record_jsonschema;
     Ppx_deriving_jsonschema_runtime.json_schema doc_comment_disabled_jsonschema;
     Ppx_deriving_jsonschema_runtime.json_schema doc_comment_override_jsonschema;